Beginning updates for new database schema
