+++ requirements.txt
@@ -1,10 +1,10 @@
-Flask==2.0.3
-SQLAlchemy==2.0.17
-PyMySQL==1.1.0
-gunicorn==20.1.0
-cloud-sql-python-connector==1.2.4
-functions-framework==3.3.0
-requests==2.27.1
-Werkzeug==2.3.7
+Flask==3.*
+SQLAlchemy
+PyMySQL
+gunicorn
+cloud-sql-python-connector
+functions-framework
+requests
+Werkzeug
 graypy
 python-dotenv
