+++ sqlconnector/connect_localconnection.py
@@ -1,7 +1,6 @@
 import os
 
 import sqlalchemy
-# import mysql
 import pymysql
 from dotenv import load_dotenv
 
@@ -15,26 +14,8 @@ hostconn=os.getenv('LOCAL_CONNECTION_IP')
 
 def local_conn():
     connection = sqlalchemy.create_engine(f"mysql+pymysql://{user}:{password}@{hostconn}/{database}")
-    # pyconnection = pymysql.connect(host="localhost",
-    #                             user=os.getenv('DB_USERNAME'),
-    #                             password=os.getenv('DB_PASSWORD'),
-    #                             database=os.getenv('DB_NAME'),
-    #                             )
 
     with connection.connect() as connect:
-        result =  connect.execute(sqlalchemy.text("SELECT * FROM characters"))
-        print(result.fetchone())
+        result =  connect.execute(sqlalchemy.text("SELECT * FROM `character`"))
 
     return connection
-
-    # try:
-    #     with pyconnection:
-    #         with pyconnection.cursor() as cursor:
-    #             cursor.execute("SELECT * from characters")
-    #             result = cursor.fetchall()
-    #             print(result)
-    # except Exception as e:
-    #     print(e)
-
-
-# [END cloud_sql_mysql_sqlalchemy_connect_connector]
