+++ sqlconnector/connection_pool.py
@@ -1,0 +1,45 @@
+
+# from __future__ import annotations
+import os
+import sqlalchemy
+
+from connect_connector import connect_with_connector
+from connect_connector_auto_iam_authn import connect_with_connector_auto_iam_authn
+from connect_tcp import connect_tcp_socket
+from connect_unix import connect_unix_socket
+from connect_localconnection import local_conn
+
+
+def init_connection_pool() -> sqlalchemy.engine.base.Engine:
+    """Sets up connection pool for the app."""
+    # use a TCP socket when INSTANCE_HOST (e.g. 127.0.0.1) is defined
+    if os.environ.get("INSTANCE_HOST"):
+        return connect_tcp_socket()
+
+    # use a Unix socket when INSTANCE_UNIX_SOCKET (e.g. /cloudsql/project:region:instance) is defined
+    if os.environ.get("INSTANCE_UNIX_SOCKET"):
+        return connect_unix_socket()
+
+    # use the connector when INSTANCE_CONNECTION_NAME (e.g. project:region:instance) is defined
+    if os.environ.get("INSTANCE_CONNECTION_NAME"):
+        # Either a DB_USER or a DB_IAM_USER should be defined. If both are
+        # defined, DB_IAM_USER takes precedence.
+        return (
+            connect_with_connector_auto_iam_authn()
+            if os.environ.get("DB_IAM_USER")
+            else connect_with_connector()
+        )
+    
+    if os.environ.get('LOCAL_CONNECTION_IP'):
+        return local_conn()
+
+    raise ValueError(
+        "Missing database connection type. Please define one of INSTANCE_HOST, INSTANCE_UNIX_SOCKET, or INSTANCE_CONNECTION_NAME"
+    )
+
+# This global variable is declared with a value of `None`, instead of calling
+# `init_db()` immediately, to simplify testing. In general, it
+# is safe to initialize your database connection pool when your script starts
+# -- there is no need to wait for the first request.
+db = init_connection_pool()
+
