+++ website/__init__.py
@@ -1,15 +1,9 @@
+import logging
 
 from __future__ import annotations
-import logging
-import os
 from flask import Flask
-import sqlalchemy
 
-from sqlconnector.connect_connector import connect_with_connector
-from sqlconnector.connect_connector_auto_iam_authn import connect_with_connector_auto_iam_authn
-from sqlconnector.connect_tcp import connect_tcp_socket
-from sqlconnector.connect_unix import connect_unix_socket
-from sqlconnector.connect_localconnection import local_conn
+from sqlconnector.connection_pool import db
 
 app = Flask(__name__, template_folder='Templates',static_folder='Static')
 app.secret_key = "123123123"
@@ -17,38 +11,11 @@ app.secret_key = "123123123"
 logger = logging.getLogger()
 
 
-def init_connection_pool() -> sqlalchemy.engine.base.Engine:
-    """Sets up connection pool for the app."""
-    # use a TCP socket when INSTANCE_HOST (e.g. 127.0.0.1) is defined
-    if os.environ.get("INSTANCE_HOST"):
-        return connect_tcp_socket()
-
-    # use a Unix socket when INSTANCE_UNIX_SOCKET (e.g. /cloudsql/project:region:instance) is defined
-    if os.environ.get("INSTANCE_UNIX_SOCKET"):
-        return connect_unix_socket()
-
-    # use the connector when INSTANCE_CONNECTION_NAME (e.g. project:region:instance) is defined
-    if os.environ.get("INSTANCE_CONNECTION_NAME"):
-        # Either a DB_USER or a DB_IAM_USER should be defined. If both are
-        # defined, DB_IAM_USER takes precedence.
-        return (
-            connect_with_connector_auto_iam_authn()
-            if os.environ.get("DB_IAM_USER")
-            else connect_with_connector()
-        )
-    
-    if os.environ.get('LOCAL_CONNECTION_IP'):
-        return local_conn()
-
-    raise ValueError(
-        "Missing database connection type. Please define one of INSTANCE_HOST, INSTANCE_UNIX_SOCKET, or INSTANCE_CONNECTION_NAME"
-    )
-
 # This global variable is declared with a value of `None`, instead of calling
 # `init_db()` immediately, to simplify testing. In general, it
 # is safe to initialize your database connection pool when your script starts
 # -- there is no need to wait for the first request.
-db = init_connection_pool()
+db = db
 
 
 from .views import views
@@ -56,4 +23,4 @@ app.register_blueprint(views, url_prefix="/")
 
 if __name__ == "__main__":
     app.run(host="127.0.0.1", port=8080, debug=True)
-    db = init_connection_pool()
+    db = db
